+++ specs/stork-deployment.yaml
@@ -72,7 +72,7 @@ rules:
   - apiGroups: [""]
     resources: ["events"]
     verbs: ["list", "watch", "create", "update", "patch"]
-  - apiGroups: ["stork.com"]
+  - apiGroups: ["stork.libopenstorage.com"]
     resources: ["storkrules"]
     verbs: ["get", "list"]
   - apiGroups: ["apiextensions.k8s.io"]
